Chain Attacks roughly implemented
Shifted my mindset from thinking about it as a true "Input Buffer" and was able to create and implement a solid "Action Chain" system that uses the character's animation frame counts to define a window of time the player can chain together further attacks. Tomorrow I plan to add the heavy attacks and a variable to store the current attack chain the player is in.