Taking advantage of the AnimationPlayer node
Set up Call Method Tracks on each of my attack animations to provide information about start-up, active, and recovery frames, as well as set a current_frame variable for the currently playing animation.
Updated the attack scripts to reflect this change. Added a heavy attack script, but the state machine loses track of the attack_queue when attempting to chain light and heavy attacks.