Way too many changes
Worked on this update for about 3 weeks. 2.5 weeks were spent trying to get local multiplayer functional to little success. During the last 4 days I took a break from that and made the beginnings of the menu system. Now there's a Main Menu that optionally opens when the game is launched, a Pause Menu, and switching between the Overworld and Fighting scenes via the menus.